% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yds_tky23.R
\name{read_yds_tky23}
\alias{read_yds_tky23}
\title{Read Yahoo Data Solution's Tokyo Special Ward File}
\usage{
read_yds_tky23(path, sheet = NULL, long = FALSE)
}
\arguments{
\item{path}{The path to the downloaded file in excel format.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet),
or an integer (the position of the sheet). It is NULL by default:
Reads all the sheets in the file (the entire period).}

\item{long}{The original data is date separated by variables,
and when \emph{TRUE} is specified, it is converted to a format that
stores the date and observable in one column each.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("superseded")}
}
\details{
Reads the "Day-to-day transition of
Tokyo's 23 ward stay population estimates" data provided by
Yahoo Data Solution in an easy-to-use format.
}
\seealso{
\url{https://ds.yahoo.co.jp/}
}
