% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yds_visitor.R
\name{read_yds_visitor}
\alias{read_yds_visitor}
\title{Read Yahoo Data Solution's File}
\usage{
read_yds_visitor(path, sheet = NULL, long = FALSE)
}
\arguments{
\item{path}{The path to the downloaded file in excel format.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet),
or an integer (the position of the sheet). It is NULL by default:
All sheets in the file except for the calendar data.}

\item{long}{The original data is date separated by variables,
and when \emph{TRUE} is specified, it is converted to a format that
stores the date and observable in one column each.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\details{
Reads the "visitor" data provided by
Yahoo Data Solution in an easy-to-use format.
}
\seealso{
\url{https://ds.yahoo.co.jp/}
}
